Refactor: Moving core ETL modules to new ETL/ directory #12
