+++ app/etl/transform.py
@@ -1,6 +1,6 @@
 import pandas as pd
 import json
-
+print("this is transform module")
 def _split_string_into_lines(historical_data: str) -> list:
     """
     Breaks a long text into a list of shorter texts (lines).
